+++ src/DotNext.Threading/Threading/AsyncLazy.cs
@@ -1,4 +1,5 @@
-﻿using System.Diagnostics;
+﻿using System.ComponentModel;
+using System.Diagnostics;
 using System.Runtime.CompilerServices;
 
 namespace DotNext.Threading;
@@ -14,17 +15,19 @@ public class AsyncLazy<T> : ISupplier<CancellationToken, Task<T>>
 {
     private const string NotAvailable = "<NotAvailable>";
     private readonly bool resettable;
-    private volatile Task<T>? task;
-
-    // null or Func<Task<T>> or Func<CancellationToken, Task<T>>
-    private MulticastDelegate? factory;
+    private readonly object syncRoot;
+    private Task<T>? task;
+    private Func<CancellationToken, Task<T>>? factory;
 
     /// <summary>
     /// Initializes a new instance of lazy value which is already computed.
     /// </summary>
     /// <param name="value">Already computed value.</param>
     public AsyncLazy(T value)
-        => task = System.Threading.Tasks.Task.FromResult(value);
+    {
+        task = System.Threading.Tasks.Task.FromResult(value);
+        syncRoot = new();
+    }
 
     /// <summary>
     /// Initializes a new instance of lazy value.
@@ -32,12 +35,10 @@ public class AsyncLazy<T> : ISupplier<CancellationToken, Task<T>>
     /// <param name="valueFactory">The function used to compute actual value.</param>
     /// <param name="resettable"><see langword="true"/> if previously computed value can be removed and computation executed again when it will be requested; <see langword="false"/> if value can be computed exactly once.</param>
     /// <exception cref="ArgumentException"><paramref name="valueFactory"/> is <see langword="null"/>.</exception>
-    [Obsolete("Use another constructor that accepts a factory with CancellationToken support.")]
+    [Obsolete("Use another constructor that accepts a factory with CancellationToken support.", error: true)]
+    [EditorBrowsable(EditorBrowsableState.Never)]
     public AsyncLazy(Func<Task<T>> valueFactory, bool resettable = false)
-    {
-        factory = valueFactory ?? throw new ArgumentNullException(nameof(valueFactory));
-        this.resettable = resettable;
-    }
+        => throw new NotImplementedException();
 
     /// <summary>
     /// Initializes a new instance of lazy value.
@@ -49,53 +50,70 @@ public class AsyncLazy<T> : ISupplier<CancellationToken, Task<T>>
     {
         factory = valueFactory ?? throw new ArgumentNullException(nameof(valueFactory));
         this.resettable = resettable;
+        syncRoot = new();
     }
 
-    private void EraseFactory()
+    private void AttachFactoryErasureCallback(Task expectedTask)
     {
-        if (task is { IsCanceled: false })
-            factory = null;
+        expectedTask.ConfigureAwait(false).GetAwaiter().UnsafeOnCompleted(EraseFactory);
+
+        void EraseFactory()
+        {
+            if (expectedTask is { IsCanceled: false } && ReferenceEquals(Volatile.Read(ref task), expectedTask))
+            {
+                lock (syncRoot)
+                {
+                    if (ReferenceEquals(task, expectedTask))
+                        factory = null;
+                }
+            }
+        }
     }
 
     /// <summary>
     /// Gets a value that indicates whether a value has been computed.
     /// </summary>
-    public bool IsValueCreated => task is { Status: TaskStatus.RanToCompletion or TaskStatus.Faulted };
+    public bool IsValueCreated => Volatile.Read(ref task) is { Status: TaskStatus.RanToCompletion or TaskStatus.Faulted };
 
     /// <summary>
     /// Gets value if it is already computed.
     /// </summary>
-    public Result<T>? Value => task.TryGetResult();
+    public Result<T>? Value => Volatile.Read(ref task).TryGetResult();
 
     /// <inheritdoc />
     Task<T> ISupplier<CancellationToken, Task<T>>.Invoke(CancellationToken token)
         => WithCancellation(token);
 
-    [MethodImpl(MethodImplOptions.Synchronized)]
     private Task<T> GetOrStartAsync(CancellationToken token)
     {
-        var t = task;
+        Task<T>? t;
+        bool fastExit;
 
-        if (t is { IsCanceled: false })
+        lock (syncRoot)
+        {
+            t = task; // read barrier is provided by monitor
+
+            if (t is { IsCanceled: false })
+            {
+                fastExit = true;
+            }
+            else
+            {
+                Debug.Assert(factory is not null);
+
+                task = t = System.Threading.Tasks.Task.Run(CreateAsyncFunc(factory, token));
+                fastExit = false;
+            }
+        }
+
+        // post-processing of task out of the lock
+        if (fastExit)
         {
             t = t.WaitAsync(token);
         }
-        else if (factory is Func<CancellationToken, Task<T>> cancelableFactory)
+        else if (!resettable)
         {
-            task = t = System.Threading.Tasks.Task.Run(CreateAsyncFunc(cancelableFactory, token));
-
-            if (!resettable)
-                t.ConfigureAwait(false).GetAwaiter().UnsafeOnCompleted(EraseFactory);
-        }
-        else
-        {
-            Debug.Assert(factory is Func<Task<T>>);
-
-            task = t = System.Threading.Tasks.Task.Run(Unsafe.As<Func<Task<T>>>(factory));
-            if (!resettable)
-                t.ConfigureAwait(false).GetAwaiter().UnsafeOnCompleted(EraseFactory);
-
-            t = t.WaitAsync(token);
+            AttachFactoryErasureCallback(t);
         }
 
         return t;
@@ -114,26 +132,32 @@ public class AsyncLazy<T> : ISupplier<CancellationToken, Task<T>>
     /// <param name="token">The token that can be used to cancel the operation.</param>
     /// <returns>Lazy representation of the value.</returns>
     public Task<T> WithCancellation(CancellationToken token)
-        => task is { IsCanceled: false } t ? t.WaitAsync(token) : GetOrStartAsync(token);
+        => Volatile.Read(ref task) is { IsCanceled: false } t ? t.WaitAsync(token) : GetOrStartAsync(token);
 
     /// <summary>
     /// Gets task representing asynchronous computation of lazy value.
     /// </summary>
     /// <seealso cref="WithCancellation"/>
     [DebuggerBrowsable(DebuggerBrowsableState.Never)]
-    [Obsolete("Use WithCancellation(CancellationToken) method instead.")]
+    [EditorBrowsable(EditorBrowsableState.Never)]
+    [Obsolete("Use WithCancellation(CancellationToken) method instead.", error: true)]
     public Task<T> Task => WithCancellation(CancellationToken.None);
 
     /// <summary>
     /// Removes already computed value from the current object.
     /// </summary>
     /// <returns><see langword="true"/> if previous value is removed successfully; <see langword="false"/> if value is still computing or this instance is not resettable.</returns>
-    [MethodImpl(MethodImplOptions.Synchronized)]
     public bool Reset()
     {
         bool result;
-        if (result = resettable && task is null or { IsCompleted: true })
-            task = null;
+        if (result = resettable && Volatile.Read(ref task) is null or { IsCompleted: true })
+        {
+            lock (syncRoot)
+            {
+                if (result = task is null or { IsCompleted: true })
+                    task = null;
+            }
+        }
 
         return result;
     }
@@ -142,7 +166,8 @@ public class AsyncLazy<T> : ISupplier<CancellationToken, Task<T>>
     /// Gets awaiter for the asynchronous operation responsible for computing value.
     /// </summary>
     /// <returns>The task awaiter.</returns>
-    [Obsolete("Use WithCancellation(CancellationToken) method instead.")]
+    [Obsolete("Use WithCancellation(CancellationToken) method instead.", error: true)]
+    [EditorBrowsable(EditorBrowsableState.Never)]
     public TaskAwaiter<T> GetAwaiter() => Task.GetAwaiter();
 
     /// <summary>
@@ -150,7 +175,8 @@ public class AsyncLazy<T> : ISupplier<CancellationToken, Task<T>>
     /// </summary>
     /// <param name="continueOnCapturedContext"><see langword="true"/> to attempt to marshal the continuation back to the original context captured; otherwise, <see langword="false"/>.</param>
     /// <returns>An object used to await asynchronous lazy initialization.</returns>
-    [Obsolete("Use WithCancellation(CancellationToken) method instead.")]
+    [Obsolete("Use WithCancellation(CancellationToken) method instead.", error: true)]
+    [EditorBrowsable(EditorBrowsableState.Never)]
     public ConfiguredTaskAwaitable<T> ConfigureAwait(bool continueOnCapturedContext)
         => Task.ConfigureAwait(continueOnCapturedContext);
 
@@ -158,10 +184,12 @@ public class AsyncLazy<T> : ISupplier<CancellationToken, Task<T>>
     /// Returns textual representation of this object.
     /// </summary>
     /// <returns>The string representing this object.</returns>
-    public override string? ToString() => task?.Status switch
+    public override string? ToString()
     {
-        null => NotAvailable,
-        TaskStatus.RanToCompletion => task.Result?.ToString(),
-        { } status => $"<{status}>",
-    };
+        return Volatile.Read(ref task) is not { } t
+            ? NotAvailable
+            : t.Status is TaskStatus.RanToCompletion
+            ? t.Result?.ToString()
+            : $"<{t.Status}>";
+    }
 }