+++ src/DotNext/Buffers/BufferWriterSlimInterpolatedStringHandler.cs
@@ -124,16 +124,12 @@ public ref struct BufferWriterSlimInterpolatedStringHandler
         }
 
         var span = buffer.Value.InternalGetSpan(alignment);
-        if (leftAlign)
-        {
-            span.Slice(value.Length, padding).Fill(Whitespace);
-            value.CopyTo(span);
-        }
-        else
-        {
-            span.Slice(0, padding).Fill(Whitespace);
-            value.CopyTo(span.Slice(padding));
-        }
+        var filler = leftAlign
+            ? span.Slice(value.Length, padding)
+            : span.TrimLength(padding, out span);
+
+        filler.Fill(Whitespace);
+        value.CopyTo(span);
 
         buffer.Value.Advance(alignment);
         count += alignment;
@@ -149,13 +145,10 @@ public ref struct BufferWriterSlimInterpolatedStringHandler
     /// </param>
     public void AppendFormatted(scoped ReadOnlySpan<char> value, int alignment)
     {
-        var leftAlign = false;
+        bool leftAlign;
 
-        if (alignment < 0)
-        {
-            leftAlign = true;
+        if (leftAlign = alignment < 0)
             alignment = -alignment;
-        }
 
         AppendFormatted(value, alignment, leftAlign);
     }
@@ -172,20 +165,17 @@ public ref struct BufferWriterSlimInterpolatedStringHandler
     /// <param name="format">The format string.</param>
     public void AppendFormatted<T>(T value, int alignment, string? format = null)
     {
-        var leftAlign = false;
+        bool leftAlign;
 
-        if (alignment < 0)
-        {
-            leftAlign = true;
+        if (leftAlign = alignment < 0)
             alignment = -alignment;
-        }
 
         switch (value)
         {
             case ISpanFormattable:
                 for (int bufferSize = alignment; ; bufferSize = bufferSize <= MaxBufferSize ? bufferSize << 1 : throw new InsufficientMemoryException())
                 {
-                    var span = buffer.Value.InternalGetSpan(bufferSize);
+                    Span<char> span = buffer.Value.InternalGetSpan(bufferSize), filler;
                     if (((ISpanFormattable)value).TryFormat(span, out var charsWritten, format, provider))
                     {
                         var padding = alignment - charsWritten;
@@ -196,12 +186,14 @@ public ref struct BufferWriterSlimInterpolatedStringHandler
                         }
                         else if (leftAlign)
                         {
-                            span.Slice(charsWritten, padding).Fill(Whitespace);
+                            filler = span.Slice(charsWritten, padding);
+                            filler.Fill(Whitespace);
                         }
                         else
                         {
-                            span.Slice(0, charsWritten).CopyTo(span.Slice(padding));
-                            span.Slice(0, padding).Fill(Whitespace);
+                            filler = span.TrimLength(padding, out var rest);
+                            span.Slice(0, charsWritten).CopyTo(rest);
+                            filler.Fill(Whitespace);
                         }
 
                         buffer.Value.Advance(alignment);