+++ src/cluster/DotNext.Net.Cluster/Net/Cluster/Consensus/Raft/CustomTransport/GenericServer.cs
@@ -43,15 +43,8 @@ internal sealed class GenericServer : Server
     private protected override MemoryOwner<byte> AllocateBuffer(int bufferSize)
         => defaultAllocator(bufferSize);
 
-    private async void HandleConnection(ConnectionContext connection)
+    private async void HandleConnection(ConnectionContext connection, int transmissionSize, EndPoint? clientAddress, MemoryAllocator<byte> allocator)
     {
-        var clientAddress = connection.RemoteEndPoint;
-
-        // determine transmission size
-        var transmissionSize = connection.Transport.Output.GetSpan().Length;
-        var allocator = connection.Features.Get<MemoryAllocator<byte>>()
-            ?? connection.Features.Get<IMemoryPoolFeature>()?.MemoryPool?.ToAllocator()
-            ?? defaultAllocator;
         var protocol = new ProtocolPipeStream(connection.Transport, allocator, transmissionSize)
         {
             WriteTimeout = (int)ReceiveTimeout.TotalMilliseconds,
@@ -111,7 +104,7 @@ internal sealed class GenericServer : Server
 
     private async Task Listen(IConnectionListener listener)
     {
-        await using (listener.ConfigureAwait(false))
+        try
         {
             while (!lifecycleToken.IsCancellationRequested && !IsDisposingOrDisposed)
             {
@@ -121,7 +114,7 @@ internal sealed class GenericServer : Server
                     if (connection is null)
                         break;
 
-                    ThreadPool.UnsafeQueueUserWorkItem(HandleConnection, connection, preferLocal: false);
+                    ThreadPool.UnsafeQueueUserWorkItem(new ConnectionHandler(this, connection), preferLocal: false);
                 }
                 catch (Exception e) when (e is ObjectDisposedException || (e is OperationCanceledException canceledEx && canceledEx.CancellationToken == lifecycleToken))
                 {
@@ -135,6 +128,10 @@ internal sealed class GenericServer : Server
             }
 
             await listener.UnbindAsync(lifecycleToken).ConfigureAwait(false);
+        }
+        finally
+        {
+            await listener.DisposeAsync().ConfigureAwait(false);
         }
     }
 
@@ -161,5 +158,30 @@ internal sealed class GenericServer : Server
     {
         Cleanup();
         return new(listenerTask ?? Task.CompletedTask);
+    }
+
+    private sealed class ConnectionHandler : Tuple<ConnectionContext, int, EndPoint?, MemoryAllocator<byte>>, IThreadPoolWorkItem
+    {
+        private readonly WeakReference<GenericServer> server;
+
+        internal ConnectionHandler(GenericServer server, ConnectionContext connection)
+            : base(connection, GetTransmissionSize(connection), connection.RemoteEndPoint, GetMemoryAllocator(server, connection))
+            => this.server = new(server, trackResurrection: false);
+
+        private static int GetTransmissionSize(ConnectionContext connection)
+            => connection.Transport.Output.GetSpan().Length;
+
+        private static MemoryAllocator<byte> GetMemoryAllocator(GenericServer server, BaseConnectionContext connection)
+        {
+            return connection.Features.Get<MemoryAllocator<byte>>()
+                ?? connection.Features.Get<IMemoryPoolFeature>()?.MemoryPool?.ToAllocator()
+                ?? server.defaultAllocator;
+        }
+
+        void IThreadPoolWorkItem.Execute()
+        {
+            if (this.server.TryGetTarget(out var server))
+                server.HandleConnection(Item1, Item2, Item3, Item4);
+        }
     }
 }